bolt | move App to domain.shared pkg
