+++ domain/shared/src/main/java/sh/kau/playground/domain/shared/App.kt
@@ -1,4 +1,4 @@
-package sh.kau.playground.domain
+package sh.kau.playground.domain.shared
 
 /**
  * This is the main application interface.
